Fix reads across chunk boundaries (#350)
* Fix reads across chunk boundaries

When using Direct-IO requests are not guaranteed to be aligned on page
size boundary, like with buffered IO. This means a read request can
start at the end of one chunk and finish at the beginning of the next
chunk. In cases where this happens we now request multiple chunks in
GetChunk and assemble the response data.

* Assemble request slices out of order

This improved performance, since we don't have to wait for each slice
sequentially, but can request them all at once.
Since the channel order does not have to match the request order, we
track a sequence number with each request to allow reassembly of the
response out of order.