Add direct_io mount option, max_readahead default
This reduces the number of FUSE read requests significantly by using
a default max_readahead value of 128 KB, which is the default readahead
used by libfuse and also the maximum request size for buffered I/O used
by the FUSE implementation in the Linux kernel.

We need to set a default, because the bazil.org/fuse package otherwise
defaults to max_readahead=0, which disables readahead completely causing
FUSE to request only 4 KB (one page) per request.

The direct_io option allows to force all files to be opened with the
O_DIRECT flag, bypassing kernel cache, I/O reordering and alignment.
Depending on the apps I/O patterns this can be benefitial, because
it avoids double-buffering of data in both the storage chunk cache
and kernel memory leading to reduced memory usage and CPU usage for
moving data in and out of kernel caches.

It will also theoretically allow FUSE requests of up to 1 MB instead
of the maximum of 128 KB, which FUSE enforces for buffered I/O, because
it splits larger requests into multiple 128 KB chunks that it requests
asynchronously.

Currently the maximum with Direct (unbuffered) I/O will still be 128 KB,
because the bazil.org/fuse package does not yet support the required
FUSE version and max_pages capability required for larger I/Os.